Enable tests that use the database
This has been a bit of a struggle because of the interaction of three
facts:
a) we need to migrate the tables before using the database
b) sqlalchemy uses a separate database connection per thread
c) pytest-flask runs the live_server in a separate process

Also, pytest-flask doesn't offer any way to run code in the process
before the app starts. So, if we wanted to keep using an in-memory
sqlite db, we'd have to a) migrate it in the app, which has implications
for prod, and b) live with the fact that we cannot access it from the
test code when using live_server.

So, we just use a file and wipe/re-init the database before every test.