Make the tests run in a well-defined environment
The new 500-test was failing on my machine since I have
FLASK_ENVIRONMENT=development set, which caused flask to re-raise the
exception instead of calling the error handler.

By explicitly setting the environment to "test", we make the behaviour
uniform across machines, but since flask also propagates exceptions in
test, we have to tell it not to.