Add support for check in with the Corona-Warn-App
The Corona-Warn-App allow checking in at location to allow warning
other users that might have been in the same location as an
infected person. For this we generate a link that is added to the
thank you page. Clicking this link will open the CWA and allow the
user to checkin for the relevant location.

To generate this link more informations are needed about the location,
like the address or type of location. We add a new class to store this
data and the code needed to generate the link.

The locations are now loaded when the application is started and we
use a dictionary indexed by the location id instead of a simple string
list. The redirection to the thank you page now include the location
ID which is then used to add the CWA link if possible. It might not be
possible to add the CWA link if some informations are missing, like
the address.